+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.1.0 (Unreleased)
 
 FEATURES:
+
+- Creates a P0 Terraform provider
+- Supports configuring P0 routing rules